+++ sinch/domains/numbers/sinch_events/v1/events/number_common_sinch_event.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 from typing import Optional
 
-from pydantic import BaseModel, Field, StrictStr
+from pydantic import Field, StrictStr
 
 from sinch.domains.numbers.sinch_events.v1.events.resource_type import (
     ResourceType,
@@ -15,4 +15,4 @@ class NumberCommonSinchEvent(SinchEvent):
     event_id: Optional[StrictStr] = Field(default=None, alias="eventId", description="The ID of the event.")
     timestamp: Optional[datetime] = Field(default=None, description="The date and time when the callback was created and added to the callbacks queue.")
     project_id: Optional[StrictStr] = Field(default=None, alias="projectId", description="The ID of the project to which the event belongs.")
-    resource_type: ResourceType = Field(default=..., alias="resourceType")
+    resource_type: ResourceType = Field(..., alias="resourceType")