+++ sinch/domains/numbers/sinch_events/v1/events/number_sinch_event.py
@@ -1,4 +1,3 @@
-from datetime import datetime
 from typing import Literal, Optional, Union
 
 from pydantic import Field, StrictStr